RGZ: fix

+++ RGZ/src/list.cpp
@@ -19,7 +19,7 @@ int list::getCount() const {
 }
 
 list::list(const list &clone) {
-    for (int i = 0; i < count; i++) {
+    for (unsigned long i = 0; i < (unsigned long) count; i++) {
         watches[i] = clone.watches[i];
     }
 }
@@ -34,16 +34,16 @@ vector<watch *> list::getLink() const {
 }
 
 watch *list::getLink(int index) {
-    return watches[index];
+    return watches[(unsigned long) index];
 }
 
 void list::clearLink(int index) {
-    if (index > watches.size()) {
-        index = watches.size();
+    if ((unsigned long) index > watches.size()) {
+        (unsigned long) index = watches.size();
     } else if (index < 0) {
-        index = 0;
+        (unsigned long) index = 0;
     }
-    auto pos = watches.begin() + index;
+    auto pos = watches.begin() + (unsigned long) index;
     watches.erase(pos);
     this->count--;
 }
@@ -57,7 +57,7 @@ void list::clearList() {
     this->count = 0;
 }
 
-void list::setLinks(const vector<watch *>& watchesCopy) {
+void list::setLinks(const vector<watch *> &watchesCopy) {
     while (!watches.empty()) {
         watches.pop_back();
     }
@@ -67,7 +67,7 @@ void list::setLinks(const vector<watch *>& watchesCopy) {
 }
 
 void list::showAll() const {
-    for (int i = 0; i < this->count; ++i) {
+    for (unsigned long i = 0; i < (unsigned long) this->count; ++i) {
         watches[i]->show();
         cout << endl;
     }
